Renamed exceptions.py and created an empty Ledger class.
